+++ lib/transpec/git.rb
@@ -3,6 +3,7 @@
 module Transpec
   module Git
     GIT = 'git'
+    COMMIT_MESSAGE_FILE_PATH = File.join('.git', 'COMMIT_EDITMSG')
 
     module_function
 
@@ -19,8 +20,23 @@ module Transpec
     end
 
     def clean?
-      fail 'The current working directory is not a Git repository' unless inside_of_repository?
+      fail_unless_inside_of_repository
       `#{GIT} status --porcelain`.empty?
+    end
+
+    def repository_root
+      fail_unless_inside_of_repository
+      `#{GIT} rev-parse --show-toplevel`.chomp
+    end
+
+    def write_commit_message(message)
+      fail_unless_inside_of_repository
+      file_path = File.join(repository_root, COMMIT_MESSAGE_FILE_PATH)
+      File.write(file_path, message)
+    end
+
+    def fail_unless_inside_of_repository
+      fail 'The current working directory is not a Git repository' unless inside_of_repository?
     end
   end
 end