+++ lib/transpec/report.rb
@@ -22,20 +22,30 @@ module Transpec
       Hash[record_counts.sort_by { |record, count| -count }]
     end
 
-    def colored_summary
+    def colored_summary(options = nil)
+      options ||= { bullet: nil, separate_by_blank_line: false }
+
+      entry_prefix = options[:bullet] ? options[:bullet] + ' ' : ''
+      indentation = if options[:bullet]
+                      ' ' * entry_prefix.length
+                    else
+                      ''
+                    end
+
       summary = ''
 
       unique_record_counts.each do |record, count|
-        summary << pluralize(count, 'conversion').color(:cyan) + "\n"
-        summary << '  ' + 'from: '.color(:cyan) + record.original_syntax + "\n"
-        summary << '    ' + 'to: '.color(:cyan) + record.converted_syntax + "\n"
+        summary << "\n" if options[:separate_by_blank_line] && !summary.empty?
+        summary << entry_prefix + pluralize(count, 'conversion').color(:cyan) + "\n"
+        summary << indentation + '  ' + 'from: '.color(:cyan) + record.original_syntax + "\n"
+        summary << indentation + '    ' + 'to: '.color(:cyan) + record.converted_syntax + "\n"
       end
 
       summary
     end
 
-    def summary
-      without_color { colored_summary }
+    def summary(options = nil)
+      without_color { colored_summary(options) }
     end
 
     def colored_stats