+++ spec/transpec/git_spec.rb
@@ -20,16 +20,6 @@ module Transpec
       end
     end
 
-    shared_context 'inside of git repository' do
-      around do |example|
-        Dir.mkdir('repo')
-        Dir.chdir('repo') do
-          `git init`
-          example.run
-        end
-      end
-    end
-
     describe '.inside_of_repository?' do
       subject { Git.inside_of_repository? }
 
@@ -97,6 +87,74 @@ module Transpec
       context 'when not inside of git repository' do
         it 'raises error' do
           -> { Git.clean? }.should raise_error(/is not a Git repository/)
+        end
+      end
+    end
+
+    describe '.repository_root' do
+      context 'when inside of git repository' do
+        include_context 'inside of git repository'
+
+        context 'and the current working directory is the repository root' do
+          it 'returns the repository root path' do
+            Git.repository_root.should == File.expand_path('.')
+          end
+        end
+
+        context 'and the current working directory is not the repository root' do
+          around do |example|
+            Dir.mkdir('dir')
+            Dir.chdir('dir') do
+              example.run
+            end
+          end
+
+          it 'returns the repository root path' do
+            Git.repository_root.should == File.expand_path('..')
+          end
+        end
+      end
+
+      context 'when not inside of git repository' do
+        it 'raises error' do
+          -> { Git.repository_root }.should raise_error(/is not a Git repository/)
+        end
+      end
+    end
+
+    describe '.write_commit_message' do
+      let(:message) { 'This is the commit message.' }
+
+      context 'when inside of git repository' do
+        include_context 'inside of git repository'
+
+        context 'and there is .git directory in the current working directory' do
+          it 'writes the message to .git/COMMIT_EDITMSG' do
+            Git.write_commit_message(message)
+            File.read('.git/COMMIT_EDITMSG').should == message
+          end
+        end
+
+        context 'and there is not .git directory in the current working directory' do
+          around do |example|
+            Dir.mkdir('dir')
+            Dir.chdir('dir') do
+              example.run
+            end
+          end
+
+          it 'writes the message to .git/COMMIT_EDITMSG in repository root directory' do
+            Git.write_commit_message(message)
+            File.read('../.git/COMMIT_EDITMSG').should == message
+          end
+        end
+      end
+
+      context 'when not inside of git repository' do
+        it 'raises error' do
+          lambda do
+            Git.write_commit_message(message)
+          end.should raise_error(/is not a Git repository/)
         end
       end
     end